Refactor error classes
