+++ src/schema/GetSchemaTask.ts
@@ -5,8 +5,8 @@ import { CfnService } from '../services/CfnService';
 import { LoggerFactory } from '../telemetry/LoggerFactory';
 import { ScopedTelemetry } from '../telemetry/ScopedTelemetry';
 import { Measure, Telemetry } from '../telemetry/TelemetryDecorator';
-import { classifyAwsError } from '../utils/AwsErrorMapper';
-import { isClientNetworkError } from '../utils/Errors';
+import { classifyAwsError } from '../utils/errors/AwsErrorMapper';
+import { isClientNetworkError } from '../utils/errors/ErrorUtils';
 import { AwsRegion } from '../utils/Region';
 import { downloadFile } from '../utils/RemoteDownload';
 import { PrivateSchemas, PrivateSchemasType, PrivateStoreKey } from './PrivateSchemas';
