+++ src/stacks/actions/ChangeSetDeletionWorkflow.ts
@@ -6,7 +6,7 @@ import { CfnInfraCore } from '../../server/CfnInfraCore';
 import { CfnService } from '../../services/CfnService';
 import { LoggerFactory } from '../../telemetry/LoggerFactory';
 import { Measure } from '../../telemetry/TelemetryDecorator';
-import { extractErrorMessage } from '../../utils/Errors';
+import { extractErrorMessage } from '../../utils/errors/ErrorUtils';
 import { toString } from '../../utils/String';
 import { processWorkflowUpdates, mapChangesToStackChanges, isStackInReview } from './StackActionOperations';
 import {
