+++ tst/unit/utils/errors/ErrorUtils.test.ts
@@ -1,7 +1,6 @@
 import { ErrorCodes, ResponseError } from 'vscode-languageserver';
 import { describe, expect, test } from 'vitest';
 import {
-    DoesNotExist,
     extractErrorCode,
     extractErrorMessage,
     extractHttpStatus,
@@ -9,9 +8,9 @@ import {
     extractStatusReason,
     handleLspError,
     isClientNetworkError,
-} from '../../../src/utils/Errors';
+} from '../../../../src/utils/errors/ErrorUtils';
 
-describe('Errors', () => {
+describe('ErrorUtils', () => {
     describe('isClientNetworkError', () => {
         test('returns true for SSL certificate errors', () => {
             expect(isClientNetworkError(new Error('unable to get local issuer certificate'))).toBe(true);
@@ -355,41 +354,6 @@ describe('Errors', () => {
             expect(extractHttpStatus('string')).toBeUndefined();
             expect(extractHttpStatus(42)).toBeUndefined();
             expect(extractHttpStatus(undefined)).toBeUndefined();
-        });
-    });
-
-    describe('DoesNotExist', () => {
-        test('formats the message with the supplied resource name', () => {
-            const error = new DoesNotExist('Stack arn');
-            expect(error.message).toBe('Stack arn does not exist');
-        });
-
-        test('sets the error name to DoesNotExist', () => {
-            expect(new DoesNotExist('thing').name).toBe('DoesNotExist');
-        });
-
-        test('survives instanceof Error', () => {
-            expect(new DoesNotExist('thing')).toBeInstanceOf(Error);
-        });
-
-        test('survives instanceof DoesNotExist (prototype chain preserved)', () => {
-            expect(new DoesNotExist('thing')).toBeInstanceOf(DoesNotExist);
-        });
-
-        test('preserves the cause from ErrorOptions', () => {
-            const cause = new Error('underlying io error');
-            const error = new DoesNotExist('Stack arn', { cause });
-            expect(error.cause).toBe(cause);
-        });
-
-        test('is throwable and catchable as an Error', () => {
-            try {
-                throw new DoesNotExist('Resource X');
-            } catch (err) {
-                expect(err).toBeInstanceOf(DoesNotExist);
-                expect(err).toBeInstanceOf(Error);
-                expect((err as Error).message).toBe('Resource X does not exist');
-            }
         });
     });
 });
